refactor: find_loaded_library
